merge after setting workflows

+++ database/migrations/2022_11_11_194651_create_doctors_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('specialization');
             $table->string('address');
             $table->string('phone');
-            $table->integer('hospital_id')->unsigned();
+            $table->bigInteger('hospital_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('hospital_id')->references('id')->on('hospitals')->onDelete('cascade');
